feat: Add street and city overlay
This commit adds the ability to display street and city information on the video overlay.

The changes include:
- Modifying `gopro_overlay/gpx.py` to parse `<street>` and `<city>` tags from GPX files.
- Creating a new `address` layout component to display the street and city information.
- Updating the XML layout to include the new `address` component.
- Adding a test for the new `address` component.
- Adding a pre-commit hook to run the tests.
