+++ gopro_overlay/layout_components/journey_map.py
@@ -1,0 +1,10 @@
+from ..widgets.map import JourneyMap
+from ..widgets.widgets import Widget
+
+
+def journey_map(at, entry, **kwargs) -> Widget:
+    return JourneyMap(
+        at=at,
+        location=lambda: entry().point,
+        **kwargs
+    )
