+++ gopro_overlay/layout_components/metric.py
@@ -2,28 +2,8 @@ from typing import Callable, TypeVar, Optional
 
 from pint import Quantity
 
-from .entry import Entry
-from .widgets.map import MovingMap, JourneyMap
-from .widgets.text import CachingText, Text
-from .widgets.widgets import Widget
-
-
-def journey_map(at, entry, **kwargs) -> Widget:
-    return JourneyMap(
-        at=at,
-        location=lambda: entry().point,
-        **kwargs
-    )
-
-
-def moving_map(at, entry, **kwargs) -> Widget:
-    return MovingMap(
-        at=at,
-        location=lambda: entry().point,
-        azimuth=lambda: entry().azi,
-        **kwargs
-    )
-
+from ..entry import Entry
+from .text import text
 
 T = TypeVar("T")
 
@@ -44,13 +24,6 @@ def metric_value(
         return default
 
     return value
-
-
-def text(cache=True, **kwargs) -> Widget:
-    if cache:
-        return CachingText(**kwargs)
-    else:
-        return Text(**kwargs)
 
 
 def metric(entry, accessor, formatter, converter=lambda x: x, cache=True, **kwargs):
