+++ gopro_overlay/layout_components/moving_map.py
@@ -1,0 +1,11 @@
+from ..widgets.map import MovingMap
+from ..widgets.widgets import Widget
+
+
+def moving_map(at, entry, **kwargs) -> Widget:
+    return MovingMap(
+        at=at,
+        location=lambda: entry().point,
+        azimuth=lambda: entry().azi,
+        **kwargs
+    )
