+++ gopro_overlay/layout_components/text.py
@@ -1,0 +1,9 @@
+from ..widgets.text import CachingText, Text
+from ..widgets.widgets import Widget
+
+
+def text(cache=True, **kwargs) -> Widget:
+    if cache:
+        return CachingText(**kwargs)
+    else:
+        return Text(**kwargs)
