+++ gopro_overlay/layouts/default-3840x2160.xml
@@ -4,6 +4,10 @@
         <component type="datetime" x="0" y="24" format="%H:%M:%S.%f" truncate="5" size="32" align="right"/>
     </composite>
 
+    <composite x="260" y="100" name="address">
+        <component type="address" x="0" y="0" size="16" align="right"/>
+    </composite>
+
     <composite x="3260" y="0" name="gps_info">
         <frame name="gps-lock" x="226" y="24" width="32" height="32" bg="0,0,0,128" cr="5" opacity="0.4">
             <component type="gps-lock-icon" size="32"/>
